Made library namespace unique.
