+++ .gitignore
@@ -10,4 +10,3 @@
 .cxx
 /.kotlin
 /wear
-/lib
