+++ app/src/main/java/com/rohankhayech/choona/view/components/TuningSelector.kt
@@ -55,9 +55,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.rohankhayech.android.util.ui.preview.ThemePreview
-import com.rohankhayech.choona.R
-import com.rohankhayech.choona.controller.tuner.Tuner
-import com.rohankhayech.choona.model.tuning.TuningEntry
+import com.rohankhayech.choona.lib.R
+import com.rohankhayech.choona.lib.controller.tuner.Tuner
+import com.rohankhayech.choona.lib.model.tuning.TuningEntry
 import com.rohankhayech.choona.view.theme.PreviewWrapper
 import com.rohankhayech.music.Tuning
