+++ app/src/main/java/com/rohankhayech/choona/view/screens/MainLayout.kt
@@ -39,10 +39,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.rohankhayech.android.util.ui.theme.m3.isLight
 import com.rohankhayech.android.util.ui.theme.m3.isTrueDark
-import com.rohankhayech.choona.model.preferences.InitialTuningType
-import com.rohankhayech.choona.model.preferences.TunerPreferences
-import com.rohankhayech.choona.model.tuning.TuningEntry
-import com.rohankhayech.choona.model.tuning.TuningList
+import com.rohankhayech.choona.lib.model.preferences.InitialTuningType
+import com.rohankhayech.choona.lib.model.preferences.TunerPreferences
+import com.rohankhayech.choona.lib.model.tuning.TuningEntry
+import com.rohankhayech.choona.lib.model.tuning.TuningList
 import com.rohankhayech.music.Tuning
 
 /**
