+++ app/src/play/java/com/rohankhayech/choona/controller/play/ReviewControllerImpl.kt
@@ -23,8 +23,8 @@ import androidx.datastore.preferences.core.edit
 import androidx.lifecycle.lifecycleScope
 import com.google.android.play.core.review.ReviewManager
 import com.google.android.play.core.review.ReviewManagerFactory
-import com.rohankhayech.choona.model.preferences.TunerPreferences
-import com.rohankhayech.choona.model.preferences.tunerPreferenceDataStore
+import com.rohankhayech.choona.lib.model.preferences.TunerPreferences
+import com.rohankhayech.choona.lib.model.preferences.tunerPreferenceDataStore
 import kotlinx.coroutines.launch
 
 /**
