--- lib/src/main/java/com/rohankhayech/choona/model/error/TunerException.kt
@@ -1,27 +1,0 @@
-/*
- * Choona - Guitar Tuner
- * Copyright (C) 2025 Rohan Khayech
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package com.rohankhayech.choona.model.error
-
-/**
- * An exception that occurs when the tuner is prevented from running.
- * @property message The detail message of the exception.
- * @property cause The cause of the exception, if any.
- * @author Rohan Khayech
- */
-class TunerException(override val message: String?, override val cause: Throwable?): Exception()
