--- lib/src/main/java/com/rohankhayech/choona/view/components/TuningMeter.kt
@@ -1,120 +1,0 @@
-/*
- * Choona - Guitar Tuner
- * Copyright (C) 2025 Rohan Khayech
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package com.rohankhayech.choona.view.components
-
-import kotlin.math.abs
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.core.tween
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.drawscope.DrawScope
-import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.unit.dp
-import com.rohankhayech.choona.controller.tuner.Tuner
-
-const val SliderInactiveTrackAlpha = 0.24f
-
-/**
- * Draws a circular meter with a variable-size indicator and background track.
- *
- * @param indicatorColor Color of the indicator.
- * @param trackColor Color of the background track. Defaults to a faded copy of the indicator color.
- * @param indicatorPosition Position of the indicator on the track, as a percentage value from -1.0 (leftmost) to 1.0 (rightmost).
- * @param indicatorSize Size of the indicator as a percentage value from 0.0 (no width) to 1.0 (full width of meter).
- */
-fun DrawScope.drawMeter(
-    indicatorColor: Color,
-    trackColor: Color = indicatorColor.copy(alpha = SliderInactiveTrackAlpha),
-    indicatorPosition: Float,
-    indicatorSize: Float,
-) {
-    // Arc size
-    val strokeWidth = 20.dp.toPx()
-    val arcSize = size.copy(height = size.height*2 - strokeWidth*2, width = size.width - strokeWidth)
-    val offset = Offset(strokeWidth/2, strokeWidth/2)
-
-    // Background Track
-    drawArc(
-        color = trackColor,
-        startAngle = -180f,
-        sweepAngle = 180f,
-        size = arcSize,
-        topLeft = offset,
-        style = Stroke(
-            width = strokeWidth,
-            cap = StrokeCap.Round
-        ),
-        useCenter = false
-    )
-
-    // Indicator
-    val startAngle = -90f
-    val indicatorSpan = indicatorSize * 180f
-    val indicatorAngle = indicatorPosition * (90f - (indicatorSpan/2)) - (indicatorSpan/2)
-    drawArc(
-        color = indicatorColor,
-        startAngle = startAngle + indicatorAngle,
-        sweepAngle = indicatorSpan,
-        size = arcSize,
-        topLeft = offset,
-        style = Stroke(
-            width = strokeWidth,
-            cap = StrokeCap.Round
-        ),
-        useCenter = false
-    )
-}
-
-@Composable
-fun animateTuningMeterIndicatorPosition(offset: Double?, showNote: Boolean): State<Float> {
-    return animateFloatAsState(
-        targetValue = remember(offset, showNote) { derivedStateOf {
-            val semitoneRange = if (showNote) 0.5 else 4
-            if (offset != null) {
-                (offset.toFloat() / semitoneRange.toFloat()).coerceIn(-1f..1f)
-            } else {
-                0f
-            }
-        }
-        }.value,
-        label = "Tuning Meter Position"
-    )
-}
-
-@Composable
-fun animateTuningMeterIndicatorSize(inTune: Boolean, onTuned: () -> Unit): State<Float> {
-    return animateFloatAsState(
-        targetValue = if (inTune) 1f else 2/180f,
-        animationSpec = if (inTune) tween(Tuner.TUNED_SUSTAIN_TIME-50, 50) else spring(),
-        finishedListener = {
-            if(it == 1f) { onTuned() }
-        },
-        label = "Tuning Indicator Size"
-    )
-}
-
-fun isInTune(offset: Double?): Boolean {
-    return offset != null && abs(offset) < Tuner.TUNED_OFFSET_THRESHOLD
-}
